requirements.txt
# Core dependencies
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
tiktoken>=0.5.1
anthropic>=0.18.0
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
numpy>=1.24.0

# File upload support
pymupdf>=1.23.0
python-docx>=1.0.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
